FEATURE: WISHLIST FRAMEWORK ADDED
This is being made so that users of Nifty Markets can add items to a wishlist or shopping cart whose data will be persisted for later use.
Feature currently adds multiple entries to wish list table but it seems like a custome query may be needed to successfully render this information upon login. Trial and error has shown
that is is not so simple to add a new wishlist item to a wishlist without garbling the json output produced from endpoint /login.
https://stackoverflow.com/questions/7308804/setting-the-correct-jpa-mapping-for-shopping-cart-items-and-product